Creation of plane class, partial work on Employee
Creating classes based on the interface discussion text file, the plane
class is how I want it, I ran out of time on the employee class and it
will have an update shortly.
I wrote out the getter classes assuming that I need to plug in some sort
of SQL interface code later.
I'm not entirely sure how much of this will be useful later, but I had
some spare time and it got me thinking about the project.