Login fully functional, view page functional
Login now saving cookies and validating check on all pages correctly.

View page functionality restored and reflects login cookie

clear cookie (logout) button functional on menu page, this could be
turned to a (LOGOUT) button on the banner of all other pages using the
Cookiehandler.clearUserPassword() method on button press

I hooked up the label in the reservation page to the login cookie so
that it pulls the user's name if logged in, however that page does not
require login as of this version so if you just go to that page it will
be blank with no name, since not logged in. This is fixed by going back
to login page manually, logging in with given user/pass, and navigating
back.